Fix OrderHistory mapping
- Replace Status object with a regular string for order details
- Change the type of totalSum from BigDecimal to Double

+++ history-order-service/src/main/java/org/projects/historyorderservice/dto/InstrumentOrderDto.java
@@ -1,11 +1,9 @@
 package org.projects.historyorderservice.dto;
 
-import java.math.BigDecimal;
-
 public record InstrumentOrderDto(
         Long id,
         String instrumentId,
-        Byte quantity,
-        BigDecimal price
+        Integer quantity,
+        Double price
 ) {
 }
