+++ history-order-service/src/main/java/org/projects/historyorderservice/dto/OrderHistoryResponseDto.java
@@ -1,12 +1,10 @@
 package org.projects.historyorderservice.dto;
 
-import java.math.BigDecimal;
-
 public record OrderHistoryResponseDto(
         String id,
         String user,
         String title,
         String status,
-        BigDecimal totalSum
+        Double totalSum
 ) {
 }
