+++ order-service/src/main/java/org/projects/orderservice/dto/InstrumentOrderDto.java
@@ -1,11 +1,9 @@
 package org.projects.orderservice.dto;
 
-import java.math.BigDecimal;
-
 public record InstrumentOrderDto(
         Long id,
         String instrumentId,
-        Byte quantity,
-        BigDecimal price
+        Integer quantity,
+        Double price
 ) {
 }
